refactor injected provider window system
