+++ src/backend/main_window.ts
@@ -86,7 +86,6 @@ export const createMainWindow = async () => {
       webviewTag: true,
       contextIsolation: true,
       nodeIntegration: true,
-      // sandbox: false,
       preload: path.join(__dirname, '../preload/preload.js'),
       webSecurity: app.isPackaged
     }
