+++ src/frontend/screens/MetaMaskHome/index.tsx
@@ -25,6 +25,7 @@ const MetaMaskHome = function ({ path = 'home.html' }: MetaMaskHomeProps) {
           )}`}
           allowpopups={trueAsStr}
           key={'metamask_home' + path + hash}
+          partition='persist:injectedProvider'
         />
       </div>
     </>
